uprise/semantic_parsing_with_constrained_lm/pyproject.toml
[tool.poetry]
name = "semantic_parsing_with_constrained_lm"
version = "0.1.0"
description = "Tools and instructions for reproducing the experiments in the paper Constrained Language Models Yield Few-Shot Semantic Parsers (EMNLP 2021)."
authors = ["Microsoft Semantic Machines <sminfo@microsoft.com>"]

[tool.poetry.dependencies]
python = "^3.7"
matplotlib = "^3.1.0"
torch = "1.6.0"
pydantic = "^1.4"
lark-parser = "^0.8.2"
requests = "^2.20.1"
cached-property = "^1.5.1"
typer = "^0.3.0"
jsons = "^0.10.1"
more_itertools = "^8.2.0"
transformers = "4.5.0"
httpx = {version = "^0.16.1", extras = ["http2"]}
datasets = "1.1.3"
appdirs = "^1.4.4"
sm-dataflow = {git = "https://github.com/microsoft/task_oriented_dialogue_as_dataflow_synthesis.git"}
blobfile = "^1.2.5"

[tool.poetry.dev-dependencies]
pytest = "^4.3.1"
black = "19.10b0"
isort = "4.3.21"
mypy = "0.782"
pylint = "2.6.0"
